==> code/01_clean_quota_sample.Rmd <==
---
title: "Data Cleaning"
author: Casey Breen
---

Summary: This script cleans the original network survey data from REACH Initiatives. These network survival data were organized into two separate files: 

* Survey dataframe: an individual-level dataframe, where each row is a survey respondents 
* Death dataframe: a death-level dataframe, where each row is a specific death 


```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))

## read in data 
survey_df <- readxl::read_xlsx(path = here("data", "data_raw", "2023-09-08", "quota_survey_2023-08-16.xlsx"), sheet = "main_networksurvival")
death_df <- readxl::read_xlsx(here("data", "data_raw", "2023-09-08", "quota_survey_2023-08-16.xlsx"), sheet = "deaths")
```


## Clean data - network survey df 

Drop missing data (and count number of  for pre-registration)

```{r}
## Number respondents who not give consent / shouldn't be in dataset (Drop, N = 68)
survey_df %>% 
  filter(consent != "yes")

## Missing number of respondents with missing household or neighbor (Drop, N = 2)
survey_df %>% 
  filter(consent == "yes") %>% 
  filter(is.na(num_total_hh_neighbour))

## Missing number of total number of kin (Drop, N = 18)
survey_df %>% 
  filter(consent == "yes") %>% 
  filter(is.na(num_total_kin))

## Missing number of total number of weight  livelihood (Drop, N = 3)
survey_df %>% 
  filter(consent == "yes") %>% 
  filter(is.na(livelihood) | is.na(education_level) | is.na(material_house) | is.na(cooking_fuel) | is.na(bed))

## Drop missing values 
survey_df_clean <- survey_df %>% 
  filter(consent == "yes") %>% 
  filter(!is.na(num_total_hh_neighbour)) %>% 
  filter(!is.na(num_total_kin)) %>% 
    filter(!is.na(livelihood) & !is.na(education_level) & !is.na(material_house) & !is.na(cooking_fuel) & !is.na(bed))

## Create new start month variable 
survey_df_clean <- survey_df_clean %>% 
  mutate(start_month = lubridate::floor_date(x = as_date(start), unit = "month"))

survey_df_clean <- survey_df_clean %>% 
  mutate(gender = ki_sex) %>% 
  mutate(age_class = case_when(
    ki_age %in% c(18:24) ~ "[18,25)",
    ki_age %in% c(25:34) ~ "[25,35)",
    ki_age %in% c(33:44) ~ "[35,45)",
    ki_age %in% c(45:54) ~ "[45,55)",
    ki_age %in% c(55:64) ~ "[55,65)",
    ki_age %in% c(65:100) ~ "[65,100]"
  )) %>% 
  mutate(health_zone = zone_de_sante_name)

# Assuming your data frame is named 'df'
survey_df_clean <- survey_df_clean %>%
  mutate(material_house_category = case_when(
    material_house %in% c("bricks", "cement", "cement_blocks", "coated_adobe") ~ "Manufactured",
    TRUE ~ "Natural / None",
  )) %>% 
  mutate(fuel_type = case_when(
    cooking_fuel %in% c("charcoal", "coal_lignite", "electricity", "kerosene") ~ "Modern",
    TRUE ~ "Traditional"
  ))

# Assuming your data frame is named 'df'
survey_df_clean <- survey_df_clean %>%
  mutate(modern_fuel_type = case_when(
    cooking_fuel %in% c("charcoal", "coal_lignite", "electricity", "kerosene") ~ 1,
    TRUE ~ 0
  )) %>% 
  mutate(manufactured_material_house  = case_when(
    material_house %in% c("bricks", "cement", "cement_blocks", "coated_adobe") ~ 1,
    TRUE ~ 0,
  )) %>% 
  mutate(bed_dummy = if_else(bed == "yes", 1, 0),
         radio_dummy = if_else(radio == "yes", 1, 0))

survey_df_clean <- survey_df_clean %>% 
  mutate(wealth_score = modern_fuel_type + manufactured_material_house + bed_dummy + radio_dummy) %>% 
  mutate(wealth_score_categorical = case_when(
    wealth_score >= 2 ~ "1", 
    TRUE ~ as.character(wealth_score)
  ))


```

## hh survey 


```{r}
survey_df_clean <- survey_df_clean %>% 
  mutate(age_u5_count = num_total_hh_under5,
  age_5_18_count = as.numeric(num_total_hh_under18), 
  age_18plus_count = num_total_hh_over18) 


survey_df_clean <- survey_df_clean %>%
  mutate(age_u5_count = case_when(
    age_u5_count >= 2 ~ "2+",
    TRUE ~ as.character(age_u5_count)),
    age_5_18_count = case_when(
      age_5_18_count >= 2 ~ "2+",
      TRUE ~ as.character(age_5_18_count)),
    age_18plus_count = case_when(
      age_18plus_count >= 2 ~ "2+",
      TRUE ~ "1" )) 

survey_df_clean <- survey_df_clean %>% 
  mutate(hh_size_categorical = case_when(
    num_total_hh %in% c(0, 1, 2, 3) ~"[0,4)",
    num_total_hh %in% c(4, 5, 6) ~"[4,7)",
    num_total_hh %in% c(7, 8, 9) ~"[7,10)",
    TRUE ~ "10+"
  )) %>% 
  mutate(hh_size_categorical_crude = case_when(
    num_total_hh %in% c(0:6) ~"[0,6)",
    TRUE ~ "7+"
  ))
```


## Clean data - death DF 

```{r}
## Drop deaths without valid date (Drop, N = 5) 
death_df_clean <- death_df %>% 
  filter(!date_join_final_death %in% c(44927, 45061, 44928, 44951, 45030))

## Rename variables and create death month variable
death_df_clean <- death_df_clean %>% 
  mutate(uuid_ki = uuid_ki4) %>% 
  mutate(health_zone = case_when(
    death_zs == 746101 ~ "Kalemie",           
    death_zs == 746601 ~ "Nyunzu",            
    death_zs == 746102 ~ "Nyemba")) %>% 
  mutate(death_month = lubridate::floor_date(x = as_date(date_join_final_death), unit = "month")) 
```

## write out files 
```{r}
## Write out csv files 
write_csv(survey_df_clean, here("data", "data_clean", "quota_sample_survey.csv"))
write_csv(death_df_clean, here("data", "data_clean", "quota_sample_deaths.csv"))
```
