==> code/04_construct_weights.Rmd <==
---
title: "Generate weights - raking"
author: "Casey Breen" 
---

Generate weights - estimation code 

```{r}
## packages 
library(tidyverse)
library(dplyr)
library(sf)
library(raster)
library(exactextractr)
library(here)
library(autumn)
library(cowplot)

## read in weighting targets 
weighting_targets <- read_csv(here("data/weighting/weighting_targets.csv"))

## read in data 
survey_df <- read_csv(here("data/clean_data/survey_df_cleaned_dec3_2023.csv")) 

## read in death df 
death_df <- read_csv(here("data/clean_data/death_df_cleaned_dec3_2023.csv"))
```

## Marginal population totals 

```{r}
# Define a function to calculate marginal percentages
calculate_marginal_percentage <- function(data, group_var) {
  data %>%
    group_by(!!sym(group_var), zone_de_sante) %>%
    summarise(count = sum(sum_value)) %>%
    mutate(covar = group_var) %>%
    rename(value = !!sym(group_var)) %>%
    group_by(zone_de_sante) %>%
    mutate(perc = count / sum(count))
}

# Apply the function for age_class and gender
age_count <- calculate_marginal_percentage(weighting_targets, "age_class")
gender_count <- calculate_marginal_percentage(weighting_targets, "gender")

# Combine the results
margin_pop <- bind_rows(age_count, gender_count) %>%
  ungroup() %>%
  dplyr::select(variable = covar, level = value, proportion = perc, zone_de_sante)
```

```{r}
## read survey df 
survey_df <- survey_df %>% 
  mutate(age_class = case_when(
    ki_age %in% c(18:24) ~ "[18,25)",
    ki_age %in% c(25:34) ~ "[25,35)",
    ki_age %in% c(33:44) ~ "[35,45)",
    ki_age %in% c(45:54) ~ "[45,55)",
    ki_age %in% c(55:64) ~ "[55,65)",
    ki_age %in% c(65:100) ~ "[65,100]"
  )) %>% 
  mutate(gender = ki_sex) %>% 
  mutate(age_gender = paste0(gender, " | ", age_class))
```

## construct weights 

```{r}
# Kalemie
margin_pop_kalemie <- margin_pop %>% 
  filter(zone_de_sante == "Kalemie") %>%
  mutate(proportion = round(proportion, 4)) %>% 
  dplyr::select(-zone_de_sante)

results_ns_kalemie <- survey_df %>% 
  filter(zone_de_sante_name == "Kalemie") %>% 
  harvest(margin_pop_kalemie)

# Nyunzu
margin_pop_nyunzu <- margin_pop %>% 
  filter(zone_de_sante == "Nyunzu") %>%
  dplyr::select(-zone_de_sante)

results_ns_nyunzu <- survey_df %>% 
  filter(zone_de_sante_name == "Nyunzu") %>% 
  harvest(margin_pop_nyunzu)

# Nyemba
margin_pop_nyemba <- margin_pop %>% 
  filter(zone_de_sante == "Nyemba") %>%
  dplyr::select(-zone_de_sante)

results_ns_nyemba <- survey_df %>% 
  filter(zone_de_sante_name == "Nyemba") %>% 
  harvest(margin_pop_nyemba)

# Combine all into one DataFrame
survey_df_weighted <- bind_rows(results_ns_kalemie, results_ns_nyunzu, results_ns_nyemba) 

## survey df weighted 
survey_df_weighted <- survey_df_weighted %>% 
  rename(weight_raking = weights)
```

## poststratification weights 

Use pre-written poststratification function 

```{r}
## poststratification weights 
survey_df_weighted <- networksurvival::generate_poststrat_weights(weighting_targets = weighting_targets, survey_df = survey_df_weighted) 
```

## weights 

```{r}
## weights 
weights <- survey_df_weighted %>% 
  dplyr::select(uuid_ki, weight_poststrat, weight_raking)

## death df 
death_weighted <- death_df %>% 
  inner_join(weights, by = c("uuid_ki"))
```


## write csv 

```{r}
## write csv 
write_csv(death_weighted, here("data/clean_data/death_df_cleaned_weighted_dec3_2023.csv"))
write_csv(survey_df_weighted, here("data/clean_data/survey_df_cleaned_weighted_dec3_2023.csv"))
```



```{r}
# ## margin comparison 
# margin_comparison <- margin_sample %>% 
#   left_join(margin_pop %>% rename(zone_de_sante_name = zone_de_sante, proportion_pop = proportion), by = c("variable", "level", "zone_de_sante_name")) 
# 
# margin_comparison %>% 
#   ggplot() +
#   geom_segment(aes(xend = proportion, x = proportion_pop, y = level, yend = level),
#                color = 'black', 
#                alpha = 0.5) + 
#   geom_point(size = 1.5, aes(x = proportion, y = level, color = "Sample")) +  # Color inside aes
#   geom_point(size = 1.5, aes(x = proportion_pop, y = level, color = "Population")) +  # Color inside aes
#   facet_grid(variable ~ zone_de_sante_name, scales = 'free_y', space = 'free_y', switch = 'y') +
#   labs(x = "Fraction of sample") +
#   scale_color_manual(name = "", values = c("Sample" = "black", 'Population' = "red")) +
#   scale_x_continuous(labels = scales::percent_format(), n.breaks = 10) +
#   theme_bw() +
#   theme(legend.position = c(1, 1), # Top-right corner
#         legend.justification = c("right", "top"), # Anchor the legend at its top-right
#         legend.direction = 'vertical',
#         legend.margin = margin(t = 0, unit = 'cm'),
#         legend.box.background = element_rect(colour = "black"),
#         legend.title = element_blank())

```


```{r}
## survey df weighted 
survey_df_weighted %>% 
  ggplot(aes(x = weight_poststrat)) + 
  geom_histogram(binwidth = 0.3, color = "black", fill = "grey") + 
  theme_cowplot() + 
  scale_x_continuous(breaks = seq(0, 9, by = 0.5), limits = c(0, 9)) # Add tick marks every 0.05 

## survey df weighted 
survey_df_weighted %>% 
  ggplot(aes(x = weight_raking)) + 
  geom_histogram(binwidth = 0.3, color = "black", fill = "grey") + 
  theme_cowplot() + 
  scale_x_continuous(breaks = seq(0, , by = 0.5), limits = c(0, 9)) # Add tick marks every 0.05 


# Assuming survey_df_weighted is your data frame with the columns weight_poststrat and weight_raking
weight_distribution_plot <- survey_df_weighted %>%
  pivot_longer(cols = c(weight_poststrat, weight_raking), 
               names_to = "weight_type", 
               values_to = "weight") %>%
  mutate(weight_type = case_when(
    weight_type == "weight_poststrat" ~ "Poststratification",
    weight_type == "weight_raking" ~ "Raking")) %>% 
  ggplot(aes(x = weight)) + 
  geom_histogram(binwidth = 0.2, color = "black", fill = "grey") + 
  facet_wrap(~weight_type, 
             scales = "free_x",
             ncol = 1) +
  theme_cowplot() + 
  scale_x_continuous(breaks = seq(0, 9, by = 0.5), limits = c(0, 9))


ggsave(plot = weight_distribution_plot, filename = here("figures/weight_distn.png"), width = 6, height = 6)
```
