==> code/04_weighting_targets_raking.Rmd <==
---
title: "Construct weighting targets for raking"
author: Casey Breen
---

Summary: Construct weighting targets using household survey


```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))
```

```{r}
## read in survey df
quota_sample_survey <- read_csv(here("data", "data_clean", "quota_sample_survey.csv"))
household_sample_respondent <- read_csv(here("data", "data_clean", "hh_sample_respondent.csv"))
household_sample_roster <- read_csv(here("data", "data_clean", "hh_sample_roster.csv")) 
```


```{r}
## roster df 
household_sample_roster <- household_sample_roster %>% 
  left_join(household_sample_respondent %>% dplyr::select(health_zone, uuid_col), 
            by = c('_submission__uuid' = 'uuid_col')) 

## roster df drop ages 
household_sample_roster_18plus <- household_sample_roster %>% 
  mutate(age_years = as.numeric(age_years)) %>% 
  filter(age_years >= 18)

## roster df gender 
household_sample_roster_18plus <- household_sample_roster_18plus %>% 
  mutate(gender = sex_roster) %>% 
   mutate(age_class = case_when(
    age_years %in% c(18:24) ~ "[18,25)",
    age_years %in% c(25:34) ~ "[25,35)",
    age_years %in% c(33:44) ~ "[35,45)",
    age_years %in% c(45:54) ~ "[45,55)",
    age_years %in% c(55:64) ~ "[55,65)",
    age_years %in% c(65:100) ~ "[65,100]"
  ))
```

```{r}
## clean up hh df 
household_sample_respondent <- household_sample_respondent %>% 
  mutate(age = as.numeric(age_repondant)) %>% 
  mutate(gender = sexe_repondant) %>% 
  mutate(education_level = case_when(
    education_level == "dontknow" ~ "none",
    TRUE ~  education_level
  )) %>% 
   mutate(age_class = case_when(
    age %in% c(18:24) ~ "[18,25)",
    age %in% c(25:34) ~ "[25,35)",
    age %in% c(33:44) ~ "[35,45)",
    age %in% c(45:54) ~ "[45,55)",
    age %in% c(55:64) ~ "[55,65)",
    age %in% c(65:100) ~ "[65,100]"
  )) %>% 
  mutate(gender = case_when(
    gender == "f'" ~ "f",
    TRUE ~ gender
  ))


# Adjust the function for recategorization with lowercase variable names and without dashes
recategorize_livelihood <- function(livelihood) {
  case_when(
    livelihood %in% c('employed_agriculture', 'self_agriculture', 'sell_anim_prod') ~ 'agriculture',
    livelihood %in% c('daily_labour_non', 'daily_labour_skilled', 'unskilled', 'skilled') ~ 'labor',
    livelihood %in% c('professional', 'clerical') ~ 'professional',
    livelihood %in% c('sales', 'trader') ~ 'sales',
    livelihood %in% c('domestic', 'other', 'not_working') ~ 'domestic_other',
    TRUE ~ 'no_response_misc' # Including 'noresponse', 'army', and any others not explicitly listed
  )
}

# Apply the recategorization to the data frames and count with variable names in lowercase
quota_sample_survey <- quota_sample_survey %>%
  mutate(livelihood = recategorize_livelihood(livelihood)) 

household_sample_respondent <- household_sample_respondent %>%
  mutate(livelihood = recategorize_livelihood(livelihood)) 
```


## estimate weighting targets using HH roster 

```{r}
# Define a function to calculate marginal percentages
calculate_marginal_percentage <- function(data, group_var) {
  data %>%
    group_by(!!sym(group_var), health_zone) %>%
    summarise(count = n()) %>%
    mutate(covar = group_var) %>%
    rename(value = !!sym(group_var)) %>%
    group_by(health_zone) %>%
    mutate(perc = count / sum(count))
}

# Apply the function for age_class and gender
age_count <- calculate_marginal_percentage(household_sample_roster_18plus, "age_class")
gender_count <- calculate_marginal_percentage(household_sample_roster_18plus, "gender")


# Assuming your data frame is named 'df'
household_sample_respondent <- household_sample_respondent %>%
  mutate(modern_fuel_type = case_when(
    cooking_fuel %in% c("charcoal", "coal_lignite", "electricity", "kerosene") ~ 1,
    TRUE ~ 0
  )) %>% 
  mutate(manufactured_material_house  = case_when(
    material_house %in% c("bricks", "cement", "cement_blocks", "coated_adobe") ~ 1,
    TRUE ~ 0,
  )) %>% 
  mutate(bed_dummy = if_else(bed == "yes", 1, 0),
         radio_dummy = if_else(radio == "yes", 1, 0))

household_sample_respondent %>% count(modern_fuel_type,
                                      manufactured_material_house,
                                      bed_dummy, 
                                      radio_dummy)

household_sample_respondent <- household_sample_respondent %>% 
  mutate(wealth_score = modern_fuel_type + manufactured_material_house + bed_dummy + radio_dummy) %>% 
  mutate(wealth_score_categorical = case_when(
    wealth_score >= 2 ~ "2+", 
    TRUE ~ as.character(wealth_score)
  ))

count_wealth_score <- calculate_marginal_percentage(household_sample_respondent, "wealth_score_categorical") %>% 
  mutate(value = as.character(value))

```


```{r}
hh_composition_targets <- household_sample_roster %>%
  filter(`_submission__uuid` != "589828a4-7d17-46a8-94a5-0d66f3cc7795") %>% 
  group_by(`_submission__uuid`, health_zone) %>% 
  summarize(hh_size = n(),
         age_u5 = sum(age_years < 5),
         age_5_18 = sum(age_years >= 5 & age_years < 18), 
         age_18plus = sum(age_years >= 18)) %>% 
  mutate(age_u5_count = case_when(
    age_u5 >= 2 ~ "2+",
    age_u5 == 1 ~ "1",
    age_u5 == 0 ~ "0"
  ),
  age_5_18_count = case_when(
    age_5_18 >= 2 ~ "2+",
    age_5_18 == 1 ~ "1",
    age_5_18 == 0 ~ "0"
  ),
  age_18plus_count = case_when(
    age_18plus >= 2 ~ "2+",
    age_18plus == 1 ~ "1",
    age_18plus == 0 ~ "0"
  )) %>% 
 select(health_zone, `_submission__uuid`, age_u5_count, age_5_18_count, age_18plus_count) %>% 
 pivot_longer(-c(health_zone, `_submission__uuid`))


hh_composition_household <- household_sample_roster %>%
  filter(`_submission__uuid` != "589828a4-7d17-46a8-94a5-0d66f3cc7795") %>% 
  group_by(`_submission__uuid`, health_zone) %>% 
  summarize(hh_size = n()) %>%
  ungroup() %>% 
  mutate(hh_size_categorical = case_when(
    hh_size %in% c(0, 1, 2, 3) ~"[0,4)",
    hh_size %in% c(4, 5, 6) ~"[4,7)",
    hh_size %in% c(7, 8, 9) ~"[7,10)",
    TRUE ~ "10+"
  )) %>% 
  count(health_zone, hh_size_categorical) %>%
  group_by(health_zone) %>%
  mutate(proportion = n / sum(n)) %>% 
  select(health_zone, level = hh_size_categorical, proportion) %>% 
  mutate(variable = "hh_size_categorical")


## hh composition 
hh_composition_targets <- hh_composition_targets %>%
    group_by(name, value, health_zone) %>%
    summarise(count = n()) %>%
    group_by(health_zone, name) %>%
    mutate(perc = count / sum(count)) %>% 
  select(health_zone, variable = name, level = value, proportion = perc)


# Combine the results
margin_pop_hh <- bind_rows(age_count, gender_count, count_wealth_score) %>%
  ungroup() %>%
  dplyr::select(variable = covar, level = value, proportion = perc, health_zone) %>% 
  bind_rows(hh_composition_targets, hh_composition_household)
```


```{r}
test <- generate_raking_weights(weighting_targets = margin_pop_hh, survey_df = quota_sample_survey)
```


save results 

```{r}
write_csv(margin_pop_hh, here("data/weighting/weighting_raking_targets.csv"))
```
