==> code/05_construct_weights_quota_sample.Rmd <==
---
title: "Generate weights - quota sample"
author: "Casey Breen" 
---

Generate weights - estimation code 

```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))

## read in weighting targets 
weighting_targets <- read_csv(here("data/weighting/weighting_targets.csv"))

## read in weighting targets 
weighting_targets_micro <- read_csv(here("data/weighting/hh_weighting_micro_df.csv"))

## read in data 
survey_df <- read_csv(here("data/data_clean/quota_sample_survey.csv")) %>%
  filter(enum != "team11")

## read in death df 
death_df <- read_csv(here("data/data_clean/quota_sample_deaths.csv")) %>%
  filter(death_enum != "team11")
```

## Marginal population totals 

```{r}
# Define a function to calculate marginal percentages
calculate_marginal_percentage <- function(data, group_var) {
  data %>%
    group_by(!!sym(group_var), health_zone) %>%
    summarise(count = sum(sum_value)) %>%
    mutate(covar = group_var) %>%
    rename(value = !!sym(group_var)) %>%
    group_by(health_zone) %>%
    mutate(perc = count / sum(count))
}

# Apply the function for age_class and gender
age_count <- calculate_marginal_percentage(weighting_targets, "age_class")
gender_count <- calculate_marginal_percentage(weighting_targets, "gender")


# Combine the results
margin_pop <- bind_rows(age_count, gender_count) %>%
  ungroup() %>%
  dplyr::select(variable = covar, level = value, proportion = perc, health_zone)
```

```{r}
## read survey df 
survey_df <- survey_df %>% 
  mutate(age_class = case_when(
    ki_age %in% c(18:24) ~ "[18,25)",
    ki_age %in% c(25:34) ~ "[25,35)",
    ki_age %in% c(33:44) ~ "[35,45)",
    ki_age %in% c(45:54) ~ "[45,55)",
    ki_age %in% c(55:64) ~ "[55,65)",
    ki_age %in% c(65:100) ~ "[65,100]"
  )) %>% 
  mutate(gender = ki_sex) %>% 
  mutate(age_gender = paste0(gender, " | ", age_class))
```


## poststratification weights 

Use pre-written poststratification function 

```{r}
## poststratification weights 
survey_df_weighted <- networksurvival::generate_poststrat_weights(weighting_targets = weighting_targets, survey_df = survey_df) 
```

## IPW weights 

```{r}
## poststratification weights 
survey_df_weighted <- networksurvival::generate_ipw_weights(weighting_targets = weighting_targets_micro, survey_df = survey_df_weighted) 
```


## weights 

```{r}
## weights 
weights <- survey_df_weighted %>% 
  dplyr::select(uuid_ki, weight_poststrat, weight_ipw)

## death df 
death_weighted <- death_df %>% 
  inner_join(weights, by = c("uuid_ki"))
```


## write csv files  

```{r}
## write csv 
write_csv(death_weighted, here("data", "data_clean", "quota_sample_deaths.csv"))
write_csv(survey_df_weighted, here("data", "data_clean", "quota_sample_survey.csv"))
```



```{r}
# Bind rows with weighting targets, for all health zones
weighting_df <- survey_df %>%
  select(health_zone, "hh_size_categorical", "hh_size_categorical_crude", "cooking_fuel",
         "material_house", "modern_fuel_type", "manufactured_material_house", "radio", "bed", "_id", "age_class", "gender", "age_u5_count", "age_5_18_count", "age_18plus_count") %>%
  mutate(inclusion = 1) %>%
  bind_rows(weighting_targets_micro %>%
              mutate(inclusion = 0))

# Initialize an empty dataframe to store results
all_zones_weighted <- data.frame()

# Unique health zones
health_zones <- unique(survey_df$health_zone)

model_results <- list()
# Loop through each health zone
for (zone in health_zones) {

  zone_df <- weighting_df %>% filter(health_zone == zone)

  # Step 1: Fit a logistic regression model for the current zone
  selection_model <- glm(inclusion ~ gender + age_class + hh_size_categorical + radio + bed + 
                           manufactured_material_house + modern_fuel_type + age_u5_count + 
                           age_5_18_count + age_18plus_count, 
                         data = zone_df, family = "binomial")

  # Store the model and extract coefficients
  model_results[[zone]] <- selection_model
}



# Generate LaTeX regression table
modelsummary(
  model_results,
  exponentiate = TRUE,  # Converts log coefficients to odds ratios
  output = here("figures", "table.tex"),  # Save as LaTeX file
  stars = TRUE,  # Add significance stars
  gof_omit = "Log.Lik",
  latex_options = "booktabs",  # Forces tabular instead of talltblr
)  # Remove unnecessary goodness-of-fit stat)


# etable::(model_results, tex = T, file = here("figures/table_reg_coefficients"), replace = T)

```
