==> code/06_maps_inset.Rmd <==
---
title: "Create Tanganyika Province Map"
author: Casey Breen
---

Summary: Create a map of the continent of Africa, with an inset of Tanganyika Province

```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))

## aires de sante 
aires_data <- sf::read_sf(here("data", "geography", "rdc_aires-de-sante"))

africa_map <- sf::read_sf(here("data", "geography", "Africa_Boundaries-shp"))
```


```{r}
# Filter data for the specific province
tanganyika_aires_data <- aires_data %>% 
  filter(Province == "Tanganyika")

# Create a new variable to identify whether each ZS should be colored or grayed out
tanganyika_aires_data <- tanganyika_aires_data %>% 
  mutate(color_group = if_else(ZS %in% c("Kalemie", "Nyemba", "Nyunzu"), as.character(ZS), "Other"))

# Group by the new variable and create a union of geometries within each group
tanganyika_aires_data_grouped <- tanganyika_aires_data %>% 
  group_by(color_group) %>% 
  summarise(geometry = st_union(geometry), .groups = "drop")


# Ensure the geometry is valid and compute the centroids
tanganyika_aires_data_grouped <- st_make_valid(tanganyika_aires_data_grouped)
centroids <- st_centroid(tanganyika_aires_data_grouped)

# Create a manual color palette: Specify your chosen colors here
my_palette <- c("Kalemie" = viridis_pal()(3)[1],
                "Nyemba" = viridis_pal()(3)[2],
                "Nyunzu" = viridis_pal()(3)[3],
                "Other" = "grey50")

# Coordinates for Kalemie Town
kalemie_town_x <- 29.2   # Longitude
kalemie_town_y <-  -5.9  # Latitude

# Create a data frame for Kalemie Town
kalemie_town_df <- data.frame(lon = kalemie_town_x, lat = kalemie_town_y)
st_geometry(kalemie_town_df) <- st_sfc(st_point(c(kalemie_town_x, kalemie_town_y)), crs = st_crs(data))
kalemie_town_df <- st_set_crs(kalemie_town_df, 4326)

# Plotting
tanganyika_prov_map <- ggplot() + 
  geom_sf(data = filter(tanganyika_aires_data, color_group != "Other"), aes(fill = color_group), alpha = 0.5, color = "black", linetype = "dotted", size = 0.2) +
  geom_sf(data = filter(tanganyika_aires_data, color_group == "Other"), aes(fill = color_group), alpha = 0.2, color = "black", linetype = "blank", size = 0.2) +
  geom_sf_text(data = filter(centroids, color_group != "Other"), aes(label = color_group), check_overlap = TRUE, size = 6 * 1.2) +
  geom_sf(data = kalemie_town_df, color = "black", shape = 18, size = 6) +  # Add the star
  geom_sf_text(data = kalemie_town_df, aes(label = "Kalemie City"), vjust = -.45, hjust = -.09, size = 7) +  # Add the label
  scale_fill_manual(values = my_palette) +
  coord_sf(datum = NA) +
  guides(fill = FALSE) +
  theme_map() # + labs(title = "Tanganyika Province, focal health zones")

## ggplot 
ggsave(plot = tanganyika_prov_map, filename = here("figures", "tanganyika_map.png"), width = 10, height = 7)
```


```{r}
# Assuming your data frame is named 'your_data_frame'
africa_map_continental <- africa_map %>%
                  filter(!NAME_0 %in% c("French Southern Territories", "Reunion", "Mayotte", "Seychelles", "Mauritius", "Cape Verde"))

# Simplify the geometry
# The Tolerance parameter controls the degree of simplification
africa_map_continental <- st_make_valid(africa_map_continental)
africa_map_continental_simplified <- st_simplify(africa_map_continental, dTolerance = 0.01, preserveTopology = TRUE)

## Make DRC map 
africa_drc_map  <- ggplot() +
 geom_sf(data = africa_map_continental_simplified, alpha = 0.2, color = "black", fill = "grey", linetype = "dotted", size = 0.2) +
  geom_sf(data = tanganyika_aires_data, alpha = 1, color = "black", fill = "blue", linetype = "blank", size = 1) +
  theme_map() + 
  geom_text(aes(x = 23, y = -2.8, label = "DRC"), color = "black", size = 6, fontface = "bold") 
  # geom_segment(aes(x = kalemie_town_x + 2, y = kalemie_town_y, xend = 70, yend = kalemie_town_y), 
  #              arrow = arrow(type = "closed", length = unit(0.2, "inches")), 
  #              color = "blue", size = 1)

## plot map 
combined_geo_plot <- cowplot::plot_grid(africa_drc_map, tanganyika_prov_map, labels = "AUTO")
  
ggsave(plot = combined_geo_plot, filename = here("figures", "africa_with_tanganyika_inset.png"), bg = "white", width = 14, height = 7)

```




```{r}
# Filter data for the specific province
tanganyika_aires_data <- aires_data %>% 
  filter(Province == "Tanganyika" | ZS == "Fizi")

# Create a new variable to identify whether each ZS should be colored or grayed out
tanganyika_aires_data <- tanganyika_aires_data %>% 
  mutate(color_group = if_else(ZS %in% c("Kalemie", "Nyemba", "Nyunzu", "Fizi"), as.character(ZS), "Other"))

# Ensure geometries are valid
tanganyika_aires_data_clean <- tanganyika_aires_data %>%
  mutate(geometry = st_make_valid(geometry))

# Optionally, you could also simplify the geometry (remove this line if not needed)
# tanganyika_aires_data_clean <- tanganyika_aires_data_clean %>%
#   mutate(geometry = st_simplify(geometry, dTolerance = 0.001))

# Now group and summarise
tanganyika_aires_data_grouped <- tanganyika_aires_data_clean %>%
  group_by(color_group) %>%
  summarise(geometry = st_union(geometry), .groups = "drop")


# Ensure the geometry is valid and compute the centroids
tanganyika_aires_data_grouped <- st_make_valid(tanganyika_aires_data_grouped)
centroids <- st_centroid(tanganyika_aires_data_grouped)


# Create a manual color palette: Specify your chosen colors here
my_palette <- c("Kalemie" = viridis_pal()(3)[1],
                "Nyemba" = viridis_pal()(3)[2],
                "Nyunzu" = viridis_pal()(3)[3],
                "Fizi" = "darkred", 
                "Other" = "grey50")

# Coordinates for Kalemie Town
kalemie_town_x <- 29.2   # Longitude
kalemie_town_y <-  -5.9  # Latitude

# Create a data frame for Kalemie Town
kalemie_town_df <- data.frame(lon = kalemie_town_x, lat = kalemie_town_y)
st_geometry(kalemie_town_df) <- st_sfc(st_point(c(kalemie_town_x, kalemie_town_y)), crs = st_crs(data))
kalemie_town_df <- st_set_crs(kalemie_town_df, 4326)

# Plotting
tanganyika_prov_map <- ggplot() + 
  geom_sf(data = filter(tanganyika_aires_data, color_group != "Other"), aes(fill = color_group), alpha = 0.5, color = "black", linetype = "dotted", size = 0.2) +
  geom_sf(data = filter(tanganyika_aires_data, color_group == "Other"), aes(fill = color_group), alpha = 0.2, color = "black", linetype = "blank", size = 0.2) +
  geom_sf_text(data = filter(centroids, color_group != "Other"), aes(label = color_group), check_overlap = TRUE, size = 6 * 1.2) +
  geom_sf(data = kalemie_town_df, color = "black", shape = 18, size = 6) +  # Add the star
  geom_sf_text(data = kalemie_town_df, aes(label = "Kalemie City"), vjust = -.45, hjust = -.09, size = 6) +  # Add the label
  scale_fill_manual(values = my_palette) +
  coord_sf(datum = NA) +
  guides(fill = FALSE) +
  theme_map() # + labs(title = "Tanganyika Province, focal health zones")

## ggplot 
ggsave(plot = tanganyika_prov_map, filename = here("figures", "tanganyika_map_with_fizi.png"), width = 6, height = 7)
```
