==> code/07_quota_sample_estimation.Rmd <==
---
title: "Estimate mortality rates"
author: Casey Breen
---

## Summary 

In this notebook, I estimate crude daily death rates using two datasets: a respondent-level dataset and a death-level dataset. I will calculate the following sets of estimates: 

* CDR pooled (Nyemba and Kalemie only, excluding Nyunzu)
* CDR disaggregated by health zone (Nyemba and Kalemie only, excluding Nyunzu)
* CDR disaggregated by month
* CDR disaggregated by month and health zone (Nyemba and Kalemie only, excluding Nyunzu)

This code calculate separate estimates for kin tie, neighbor tie, and blended. It also by default calculates both weighted (post-stratification survey weights constructed using targets from worldpop) and unweighted estimates by default. 

The uncertainty intervals will come from 10,000 bootstrap resamples, where re-estimate the quantity of interest and take the 2.5 and 97.5th percentile. And to estimate uncertainty intervals for the weighted estimates, the code reconstructs survey weights for each bootstrap sample.

```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))

## read in survey df
quota_sample_survey <- read_csv(here("data", "data_clean", "quota_sample_survey.csv")) # %>% filter(health_zone != "Nyunzu")

quota_sample_survey <- quota_sample_survey %>%
  filter(!(start_month == as_date("2023-06-01") & health_zone == "Nyunzu"))

## read in death df 
quota_sample_deaths <- read_csv(here("data", "data_clean", "quota_sample_deaths.csv")) # %>%  filter(health_zone != "Nyunzu")

quota_sample_deaths <- quota_sample_deaths %>% mutate(report_month = floor_date(death_date_dc, "month")) %>% filter(!(report_month == as_date("2023-06-01") & health_zone == "Nyunzu"))

## read in weighting targets 
weighting_targets_poststrat <- read_csv(here("data", "weighting", "weighting_targets.csv")) 

## read in weighting targets 
weighting_targets_raking <- read_csv(here("data", "weighting", "weighting_raking_targets.csv"))

## hh weighting micro df 
weight_targets_ipw <- read_csv(here("data/weighting/hh_weighting_micro_df.csv"))

## set seed
set.seed(1)
```


## CDR pooled (Nyemba and Kalemie only, excluding Nyunzu)

The code below produces estimates of the pooled cdr with corresponding 95% uncertainty intervals 

```{r}
## calculate cdr poststratification  
cdr_pooled_point_poststrat <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                  survey_df = quota_sample_survey,
                                                                  weight_col = "weight_poststrat")


## calculate cdr poststratification  
cdr_pooled_bootstrap_poststrat <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                   survey_df = quota_sample_survey,
                                                                   bootstrap = 100,
                                                                   weight_col = "weight_poststrat",
                                                                   weight_type = "poststrat",
                                                                   weight_targets = weighting_targets_poststrat)

## calculate cdr ipw  
cdr_pooled_point_ipw <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                  survey_df = quota_sample_survey,
                                                                  weight_col = "weight_ipw")


## calculate cdr ipw  
cdr_pooled_bootstrap_ipw <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,       
                                                                   survey_df = quota_sample_survey, 
                                                                   bootstrap = 100,
                                                                   weight_col = "weight_ipw",
                                                                   weight_type = "ipw",
                                                                   weight_targets = weight_targets_ipw)

## uncertainty estimates 
cdr_pooled_uncertainty_ipw <- cdr_pooled_bootstrap_ipw %>% 
  group_by(type, weights) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025),
            death_rate_upper = quantile(death_rate, 0.975))

## combine point estimates and 
cdr_pooled_point_uncertainty_ipw <- inner_join(cdr_pooled_point_ipw, cdr_pooled_uncertainty_ipw, by = c("weights", "type")) %>% 
  filter(weights == "poststrat") %>% 
  mutate(weights = "ipw")


## uncertainty estimates 
cdr_pooled_uncertainty_poststrat <- cdr_pooled_bootstrap_poststrat %>% 
  group_by(type, weights) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025),
            death_rate_upper = quantile(death_rate, 0.975))

cdr_pooled_point_uncertainty_poststrat <- inner_join(cdr_pooled_point_poststrat, cdr_pooled_uncertainty_poststrat, by = c("weights", "type"))

# Create visualizations 
cdr_pooled_plot <- cdr_pooled_point_uncertainty_ipw %>% 
  bind_rows(cdr_pooled_point_uncertainty_poststrat) %>%
  filter(type != "household") %>% 
  mutate(
    type = factor(type, levels = c("kin", "neighbor", "blended")),
    weights = factor(case_when(
      weights == "poststrat" ~ "Poststratification",
      weights == "unweighted" ~ "Unweighted",
      weights == "ipw" ~ "Inverse Probability"
    ), levels = c("Unweighted", "Poststratification", "Inverse Probability"))
  ) %>%
  ggplot(aes(x = type, color = weights, y = death_rate, shape = weights, ymin = death_rate_lower, ymax = death_rate_upper)) + 
  geom_pointrange(position = position_dodge(.2), size = .75, fill = "white") +
  scale_shape_manual(values = c(21, 22, 23, 24)) +
  ggsci::scale_color_lancet() + 
  scale_color_manual(values = cud) + 
  ylim(0, 1) + 
  theme_half_open(12) + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Network Tie", y = "Crude Death Rate \n (per 10,000 person days)")


# Create visualizations 
cdr_pooled_plot_unweighted <- cdr_pooled_point_uncertainty_ipw %>% 
  bind_rows(cdr_pooled_point_uncertainty_poststrat) %>%
  mutate(death_rate = ifelse(weights == "unweighted", death_rate, -999)) %>% 
  filter(type != "household") %>% 
  mutate(
    type = factor(type, levels = c("kin", "neighbor", "blended")),
    weights = factor(case_when(
      weights == "poststrat" ~ "Poststratification",
      weights == "unweighted" ~ "Unweighted",
      weights == "ipw" ~ "Inverse Probability"
    ), levels = c("Unweighted", "Poststratification", "Inverse Probability"))
  ) %>%
  ggplot(aes(x = type, color = weights, y = death_rate, shape = weights, ymin = death_rate_lower, ymax = death_rate_upper)) + 
  geom_pointrange(position = position_dodge(.2), size = .75, fill = "white") +
  scale_shape_manual(values = c(21, 22, 23, 24)) +
  ggsci::scale_color_lancet() + 
  scale_color_manual(values = cud) + 
  ylim(0, 1) + 
  theme_half_open(12) + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Network Tie", y = "Crude Death Rate \n (per 10,000 person days)")


# Create visualizations 
cdr_pooled_plot_poststrat <- cdr_pooled_point_uncertainty_ipw %>% 
  bind_rows(cdr_pooled_point_uncertainty_poststrat) %>%
  mutate(death_rate = ifelse(weights %in% c("unweighted", "poststrat"), death_rate, -999)) %>% 
  filter(type != "household") %>% 
  mutate(
    type = factor(type, levels = c("kin", "neighbor", "blended")),
    weights = factor(case_when(
      weights == "poststrat" ~ "Poststratification",
      weights == "unweighted" ~ "Unweighted",
      weights == "ipw" ~ "Inverse Probability"
    ), levels = c("Unweighted", "Poststratification", "Inverse Probability"))
  ) %>%
  ggplot(aes(x = type, color = weights, y = death_rate, shape = weights, ymin = death_rate_lower, ymax = death_rate_upper)) + 
  geom_pointrange(position = position_dodge(.2), size = .75, fill = "white") +
  scale_shape_manual(values = c(21, 22, 23, 24)) +
  ggsci::scale_color_lancet() + 
  scale_color_manual(values = cud) + 
  ylim(0, 1) + 
  theme_half_open(12) + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Network Tie", y = "Crude Death Rate \n (per 10,000 person days)")


## save main estimate plot 
ggsave(plot = cdr_pooled_plot, filename = here("figures", "cdr_pooled_plot.png"), width = 6, height = 4)
ggsave(plot = cdr_pooled_plot_unweighted, filename = here("figures", "cdr_pooled_plot_unweighted.png"), width = 6, height = 4)
ggsave(plot = cdr_pooled_plot_poststrat, filename = here("figures", "cdr_pooled_plot_poststrat.png"), width = 6, height = 4)

```



## CDR disaggregated by health zone (Nyemba and Kalemie only, excluding Nyunzu)

```{r}
## estimate point estimates 
cdr_pooled_hz_point <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                  survey_df = quota_sample_survey,
                                                                  weight_col = "weight_ipw",
                                                                  subpopulation = "health_zone",
                                                                  bootstrap = NA)

cdr_pooled_hz_boostrap <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                     survey_df = quota_sample_survey,
                                                                     weight_col = "weight_ipw",
                                                                     subpopulation = "health_zone", 
                                                                     weight_targets = weight_targets_ipw,
                                                                     weight_type = "ipw",
                                                                     bootstrap = 100)

## zone de sante - Uncertainty Intervals
cdr_pooled_hz_uncertainty <- cdr_pooled_hz_boostrap %>% 
  group_by(health_zone, type, weights) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025),
            death_rate_upper = quantile(death_rate, 0.975)) 

## perform join 
cdr_pooled_hz_point_uncertainty <- inner_join(cdr_pooled_hz_point, cdr_pooled_hz_uncertainty, by = c("type", "health_zone", "weights"))

## visualize results 
cdr_pooled_heath_zones_plot <- cdr_pooled_hz_point_uncertainty %>% 
  filter(weights == "poststrat") %>% 
  filter(type != "household") %>% 
  ggplot(aes(x = health_zone, color = type, shape = type, y = death_rate, ymin = death_rate_lower, ymax = death_rate_upper)) + 
  geom_pointrange( position = position_dodge(.25), size = 1, fill = "white") +
  scale_shape_manual(values = c(21, 22, 23)) +
  scale_color_viridis_d() +  
  ylim(0, 1) + 
  theme_cowplot() + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Health Zone", y = "Crude Death Rate \n (per 10,000 person days)")

ggsave(plot = cdr_pooled_heath_zones_plot, filename = here("figures", "cdr_pooled_heath_zones_plot.png"), width = 7, height = 5)
```



```{r}
## estimate point estimates 
cdr_pooled_hz_point <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                  survey_df = quota_sample_survey,
                                                                  weight_col = "weight_ipw",
                                                                  subpopulation = "health_zone",
                                                                  bootstrap = NA)
```


## CDR disaggregated by month

```{r}
## estimate point estimates 
cdr_monthly_point <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                survey_df = quota_sample_survey,
                                                                weight_col = "weight_ipw",
                                                                bootstrap = NA,
                                                                monthly = T)

## cdr monthly boostrap estimates 
cdr_monthly_bootstrap <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                    survey_df = quota_sample_survey,
                                                                    weight_col = "weight_ipw",
                                                                    weight_targets = weight_targets_ipw,
                                                                    weight_type = "ipw",
                                                                    bootstrap = 300,
                                                                    monthly = T)

## zone de sante 
cdr_monthly_uncertainty <- cdr_monthly_bootstrap %>% 
  filter(type != "household") %>% 
  group_by(month, type) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025),
            death_rate_upper = quantile(death_rate, 0.975))

## join together point and uncertainty estimate 
cdr_monthly_point_uncertainty <- inner_join(cdr_monthly_uncertainty, cdr_monthly_point, by = c("type", "month"))

## visualize cdr point esitmates 
cdr_monthly_plot <- cdr_monthly_point_uncertainty %>% 
  filter(type == "blended" & weights == "poststrat") %>% 
  ggplot(aes(x = month, y = death_rate)) + 
  geom_ribbon(aes(ymin = death_rate_lower, ymax = death_rate_upper), alpha = 0.1, color = "black", fill = "blue", position = position_dodge(0)) +
    geom_line(linetype = "dashed", alpha = 0.3, color = "black") +
  scale_shape_manual(values = c(21, 22, 23)) +
  ggsci::scale_color_lancet() +
  theme_cowplot() + 
  ylim(0, 1) +
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Month of Death", y = "Crude Death Rate")

ggsave(plot = cdr_monthly_plot, filename = here("figures", "cdr_monthly_plot.png"), width = 6, height = 4)
```






## CDR disaggregated by month and health zone (Nyemba and Kalemie only, excluding Nyunzu)

```{r}
## estimate point estimates 
cdr_monthly_hz_point <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                   survey_df = quota_sample_survey, 
                                                                   weight_col = "weight_ipw",
                                                                   subpopulation = "health_zone",
                                                                   bootstrap = NA,
                                                                   monthly = T)

cdr_monthly_hz_bootstrap <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                       survey_df = quota_sample_survey, 
                                                                       weight_col = "weight_ipw",
                                                                       weight_type = "ipw", 
                                                                       weight_targets = weight_targets_ipw,
                                                                       subpopulation = "health_zone", 
                                                                       bootstrap = 100, 
                                                                       monthly = T)

## health_zone
cdr_monthly_hz_uncertainty <- cdr_monthly_hz_bootstrap %>% 
  group_by(month, health_zone, type) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025, na.rm = T),
            death_rate_upper = quantile(death_rate, 0.975, na.rm = T))

## overall estimates 
cdr_monthly_hz_point_uncertainty <- inner_join(cdr_monthly_hz_uncertainty, cdr_monthly_hz_point, by = c("type", "health_zone", "month"))

## results monthly df 
cdr_monthly_hr_plot <- cdr_monthly_hz_point_uncertainty %>% 
 filter(month != as_date("2023-06-01")) %>% 
  filter(weights == "poststrat") %>% 
  filter(health_zone != "Nyemba") %>% 
  ggplot(aes(x = month, color = type, shape = type, y = death_rate, ymin = death_rate_lower, ymax = death_rate_upper)) + 
  geom_pointrange( position = position_dodge(5), size = .75, fill = "white") +
  geom_line(linetype = "dashed", alpha = 0.3) + 
  scale_shape_manual(values = c(21, 22, 23)) +
  scale_color_viridis_d() + 
  theme_cowplot() + 
  ylim(0, 2) + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Month of Death", y = "Estimate Average Daily Crude Death Rate") + 
  facet_wrap(~health_zone, ncol = 1)

ggsave(plot = cdr_monthly_hr_plot, filename = here("figures", "cdr_monthly_hz_plot.png"), width = 6, height = 8)
```
