==> code/09_hh_sample_network_estimation.Rmd <==
---
title: "Estimate mortality network method"
author: Casey Breen
---

```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))
 
## read in datasets
hh_sample_mortality_long <- read_csv(here("data", "data_clean", "hh_sample_mortality_long.csv"))
hh_sample_deaths <- read_csv(here("data", "data_clean", "hh_sample_deaths.csv"))
hh_sample_neighbors <- read_csv(here("data", "data_clean", "hh_sample_neighbors.csv"))
hh_sample_respondent <- read_csv(here("data", "data_clean", "hh_sample_respondent.csv")) 

## read in survey df
## read in survey df
quota_sample_survey <- read_csv(here("data", "data_clean", "quota_sample_survey.csv")) # %>% filter(health_zone != "Nyunzu")
quota_sample_survey <- quota_sample_survey %>%
  filter(!(start_month == as_date("2023-06-01") & health_zone == "Nyunzu"))

## read in death df 
quota_sample_deaths <- read_csv(here("data", "data_clean", "quota_sample_deaths.csv")) # %>%  filter(health_zone != "Nyunzu")
quota_sample_deaths <- quota_sample_deaths %>% mutate(report_month = floor_date(death_date_dc, "month")) %>% filter(!(report_month == as_date("2023-06-01") & health_zone == "Nyunzu"))


## read in weighting targets 
weighting_targets <- read_csv(here("data", "weighting", "weighting_targets.csv"))

## read in weighting targets 
weighting_targets_raking <- read_csv(here("data", "weighting", "weighting_raking_targets.csv"))

weight_targets_ipw <- read_csv(here("data/weighting/hh_weighting_micro_df.csv"))
```



```{r}
hh_sample_respondent <- hh_sample_respondent %>% 
  mutate(health_zone = health_zone,
         gender = case_when(
           sexe_repondant == "f" ~ "f",
           sexe_repondant == "m" ~ "m",
           sexe_repondant == "f'" ~ "f"
          ),
         age_class = case_when(
    age_repondant %in% c(18:24) ~ "[18,25)",
    age_repondant %in% c(25:34) ~ "[25,35)",
    age_repondant %in% c(33:44) ~ "[35,45)",
    age_repondant %in% c(45:54) ~ "[45,55)",
    age_repondant %in% c(55:64) ~ "[55,65)",
    age_repondant %in% c(65:100) ~ "[65,100]"
  )) %>% 
  mutate(age_gender = paste0(gender, " | ", age_class))


hh_sample_respondent <- hh_sample_respondent %>% 
  mutate(num_total_kin = coalesce(a_total_kin, b_total_kin)) %>% 
  mutate(num_total_kin = as.numeric(num_total_kin)) %>% 
  mutate(sum_neighbours = coalesce(a_sum_neighbours, b_sum_neighbours)) %>% 
  mutate(num_total_hh_neighbour = as.numeric(sum_neighbours) + as.numeric(num_hh)) %>% 
  mutate(num_total_neighbour = as.numeric(sum_neighbours)) %>% 
  mutate(num_total_hh = as.numeric(num_hh)) %>% 
  mutate(start_month = 1,
         weight_poststrat = 1) %>% 
  mutate(uuid_ki = uuid_col)

hh_sample_deaths <- hh_sample_deaths %>%  
  mutate(`death_relationship/family` = is_kin,
         `death_relationship/neighbour` = is_neighbour,
         `death_relationship/household` = is_hh) %>% 
  mutate(uuid_ki = `_submission__uuid`) %>% 
  mutate(weight_poststrat = 1)
```


```{r}
## estimate point estimates 
cdr_pooled_point_nonprob <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                          survey_df = quota_sample_survey,
                                                                          weight_col = "weight_ipw",
                                                                          bootstrap = NA)

cdr_pooled_boostrap_nonprob <- networksurvival::compute_cdr_comprehensive(death_df = quota_sample_deaths,
                                                                             survey_df = quota_sample_survey, 
                                                                             weight_col = "weight_ipw",
                                                                             weight_targets = weight_targets_ipw,
                                                                             weight_type = "ipw", 
                                                                             bootstrap = 300)

## zone de sante - Uncertainty Intervals
cdr_pooled_uncertainty_nonprob <- cdr_pooled_boostrap_nonprob %>% 
  group_by(type, weights) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025),
            death_rate_upper = quantile(death_rate, 0.975)) 

## perform join 
cdr_pooled_point_uncertainty_nonprob <- inner_join(cdr_pooled_point_nonprob, cdr_pooled_uncertainty_nonprob, by = c("type", "weights")) %>% 
  mutate(survey = "quota")
```



```{r}
## estimate point estimates 
cdr_pooled_point_hh <- networksurvival::compute_cdr_comprehensive(death_df = hh_sample_deaths,
                                                                          survey_df = hh_sample_respondent,
                                                                          weight_col = "weight_poststrat",
                                                                          bootstrap = NA)

cdr_pooled_boostrap_hh <- networksurvival::compute_cdr_comprehensive(death_df = hh_sample_deaths,
                                                                             survey_df = hh_sample_respondent, 
                                                                          weight_col = "weight_poststrat",
                                                                             bootstrap = 300)


## zone de sante - Uncertainty Intervals
cdr_pooled_uncertainty_hh <- cdr_pooled_boostrap_hh %>% 
  group_by(type, weights) %>% 
  summarize(death_rate_lower = quantile(death_rate, 0.025),
            death_rate_upper = quantile(death_rate, 0.975)) 

## perform join 
cdr_pooled_point_uncertainty_hh <- inner_join(cdr_pooled_point_hh, cdr_pooled_uncertainty_hh, by = c("type", "weights")) %>% 
  mutate(survey = "hh")


un_estimate <- 10 * 10/365 
smart_survey <- 0.58

## visualize results 
cdr_pooled_heath_zones_plot <- cdr_pooled_point_uncertainty_hh %>% 
  bind_rows(cdr_pooled_point_uncertainty_nonprob) %>% 
  filter(weights == "poststrat") %>% 
#   filter(type != "household") %>% 
  ggplot(aes(x = type, shape = survey, color = survey, y = death_rate, ymin = death_rate_lower, ymax = death_rate_upper)) + 
  geom_pointrange( position = position_dodge(.25), size = 1, fill = "white") +
  scale_shape_manual(values = c(21, 22)) +
  ylim(0, 1.2) + 
  theme_cowplot() + 
  ggsci::scale_color_lancet() + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Zone de Sante",
       y = "Estimated Average Daily Crude Death Rate",
       labs = "HH Survey Estimates") + 
  geom_hline(yintercept = un_estimate, linetype = "dashed", color = "grey")

ggsave(plot = cdr_pooled_heath_zones_plot, filename = here("figures", "cdr_pooled_heath_zones_plot.png"), width = 8, height = 5)
```





```{r}
household_wrong <- 32.68 * 10/365
household_adjusted <- 9.2 * 10/365
household_adjusted_lower <- 5.4 * 10/365
household_adjusted_upper <- 12.9 * 10/365


auxiliary_estimates <- tibble(
  survey = c("Probability", "Gold Standard", "Probability"),
  type = c("Jarrett et al. (2020)", "Jarrett et al. (2020)", "Smart Survey 2022"), 
  death_rate = c(household_wrong, household_adjusted, 0.58),
  death_rate_upper = c(0.7711899, household_adjusted_lower, 0.34), 
  death_rate_lower = c(1.019499, household_adjusted_upper, 0.98)
)

# Assuming 'un_estimate' is defined somewhere in your code as the UN estimate value.
cdr_comparison_external <- cdr_pooled_point_uncertainty_hh %>%
  bind_rows(cdr_pooled_point_uncertainty_nonprob) %>%
  mutate(survey = case_when(
    survey == "hh" ~ "Probability",
    survey == "quota" ~ "Quota"
  )) %>%
  filter(weights == "poststrat") %>%
  mutate(type = case_when(
    type == "blended" ~ "Network Survival",
    type == "household" ~ "Household",
    TRUE ~ NA_character_
  )) %>%
  filter(!is.na(type)) %>%
  bind_rows(auxiliary_estimates) %>%
  mutate(
    type = factor(type, levels = c("Network Survival", "Household", "Jarrett et al. (2020)", "Smart Survey 2022")),
    study_category = ifelse(type %in% c("Network Survival", "Household"), "This Study", "Other Studies"),
    un_label = ifelse(type %in% c("Network Survival", "Household"), "DRC National-Level Estimate (UN)", NA)
  ) %>%
  mutate(study_category = factor(study_category, levels = c("This Study", "Other Studies"))) %>%
  ggplot(aes(x = type, shape = survey, color = survey, y = death_rate, ymin = death_rate_lower, ymax = death_rate_upper)) +
  geom_pointrange(position = position_dodge(width = 0.25), size = 1, fill = "white") +
  scale_shape_manual(values = c("Probability" = 21,   # Circle for Probability
                                "Quota" = 23,        # Diamond for Quota
                                "Gold Standard" = 22  # Square for gold standard 
  )) +
  scale_color_manual(values = c("Probability" = "#0072B2",  # Blue for Probability
                                "Quota" = "#009E73",        # Green for Quota
                                "Gold Standard" = "#E69F00"  # Orange for gold standard
  )) +
  ylim(0, 1.03) +
  theme_cowplot() +
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    strip.background = element_blank(),
    strip.text.x = element_text(size = 12, colour = "black"),
    panel.spacing.x = unit(0, "lines")  # Reduced spacing between panels
  ) +
  labs(x = "", y = "Crude Death Rate \n (per 10,000 person-days)") +
  geom_hline(yintercept = un_estimate, linetype = "dashed", color = "grey") +  # DRC national-level estimate line for both panels
  facet_wrap(~study_category, scales = "free_x", ncol = 2) 

# Assuming 'un_estimate' is defined somewhere in your code as the UN estimate value.
cdr_comparison_internal <- cdr_pooled_point_uncertainty_hh %>%
  bind_rows(cdr_pooled_point_uncertainty_nonprob) %>%
  mutate(survey = case_when(
    survey == "hh" ~ "Probability",
    survey == "quota" ~ "Quota"
  )) %>%
  filter(weights == "poststrat") %>%
  mutate(type = case_when(
    type == "blended" ~ "Network Survival",
    type == "household" ~ "Household",
    TRUE ~ NA_character_
  )) %>%
  filter(!is.na(type)) %>%
  bind_rows(auxiliary_estimates) %>%
  mutate(
    type = factor(type, levels = c("Network Survival", "Household", "Jarrett et al. (2020)", "Smart Survey 2022")),
    study_category = ifelse(type %in% c("Network Survival", "Household"), "This Study", "External Studies"),
    un_label = ifelse(type %in% c("Network Survival", "Household"), "DRC National-Level Estimate (UN)", NA)
  ) %>%
  mutate(study_category = factor(study_category, levels = c("This Study", "External Studies"))) %>%
  mutate(death_rate = ifelse(study_category == "External Studies", -999, death_rate),
         death_rate_lower = ifelse(study_category == "External Studies", -999, death_rate_lower),
         death_rate_upper = ifelse(study_category == "External Studies", -999, death_rate_upper)) %>% 
  ggplot(aes(x = type, shape = survey, color = survey, y = death_rate, ymin = death_rate_lower, ymax = death_rate_upper)) +
  geom_pointrange(position = position_dodge(width = 0.25), size = 1, fill = "white") +
  scale_shape_manual(values = c(21, 22, 23)) +
  scale_color_manual(values = c("Probability" = "#0072B2",  # Blue for Probability
                                "Quota" = "#009E73",        # Green for Quota
                                "Another Quota" = "#E69F00"  # Orange for a third category, if applicable
                                )) +
  ylim(0, 1.03) +
  theme_cowplot() +
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    strip.background = element_blank(),
    strip.text.x = element_text(size = 12, colour = "black"),
    panel.spacing.x = unit(0, "lines")  # Reduced spacing between panels
  ) +
  labs(x = "", y = "Crude Death Rate \n (per 10,000 person-days)") +
  geom_hline(yintercept = un_estimate, linetype = "dashed", color = "grey") +  # DRC national-level estimate line for both panels
  facet_wrap(~study_category, scales = "free_x", ncol = 2) 


# Finally, it is possible to move the annotation inside of the plot if desired.
cdr_comparison_external <- ggdraw(add_sub(cdr_comparison_external, "DRC National-Level Estimate (UN)", vpadding=grid::unit(0, "lines"),
       y = 16, x = 0.03, hjust = 0, color = "grey36", size = 11))

ggsave(plot = cdr_comparison_external, filename = here("figures", "cdr_comparison_external.png"), width = 8, height = 5)
```

## weighting surveys 

```{r}
margin <- weight_targets_ipw %>% 
  mutate(survey = "Household") %>% 
  bind_rows(quota_sample_survey %>% 
              mutate(survey = "quota")) %>% 
  select(survey, gender, age_class, hh_size_categorical, radio, bed, manufactured_material_house, modern_fuel_type, age_u5_count, age_5_18_count,  age_18plus_count, health_zone)
```


```{r}
margin_df <- margin %>%
  mutate(manufactured_material_house = case_when(
         manufactured_material_house == 1 ~ "Modern",
         manufactured_material_house == 0 ~ "Traditional")) %>% 
  mutate(modern_fuel_type = case_when(
    modern_fuel_type == 1 ~ "Modern",
    TRUE ~ "Traditional"
  )) %>% 
  mutate(radio = case_when(
    radio == "yes" ~ "Yes",
    TRUE ~ "No"
  )) %>% 
   mutate(bed = case_when(
    bed == "yes" ~ "Yes",
    TRUE ~ "No"
  )) %>% 
    mutate(gender = case_when(
    gender == "m" ~ "Male",
    TRUE ~ "Female"
  )) %>% 
  pivot_longer(-c(survey, health_zone)) %>% 
  group_by(survey, name, health_zone) %>% 
  count(value) %>% 
  group_by(survey, name, health_zone) %>% 
  mutate(n = n / sum(n))

margin_df <- margin_df %>% 
  pivot_wider(names_from = survey, values_from = n) %>% 
  select(variable = name, level = value, proportion = quota, proportion_pop = Household) 
```




```{r}
margin_comparison_plot <- margin_df %>%  
  filter(!is.na(level)) %>% 
  mutate(variable = case_when( 
    variable == "age_18plus_count" ~ "HH 18+", 
    variable == "age_5_18_count"   ~ "HH 5-18", 
    variable == "age_u5_count" ~ "HH U5", 
    variable == "hh_size_categorical" ~ "HH Size", 
    variable == "wealth_score_categorical" ~ "Wealth Score",
    variable == "gender" ~ "Gender", 
    variable == "age_class" ~ "Age",
    variable == "bed" ~ "Bed",
    variable == "radio" ~ "Radio",
    variable == "manufactured_material_house" ~ "Wall type",
    variable == "modern_fuel_type" ~ "Fuel Type"
  )) %>% 
  # filter(health_zone != "Nyemba") %>%
  filter(!variable %in% c("HH 18+", "HH 5-18", "HH U5")) %>%
  mutate(variable = factor(variable, levels = c(
    "Age", "Gender", "HH Size", "Bed", "Fuel Type", "Radio", "Wall type"
  ))) %>% # Adjust the factor levels here
  ggplot() +
  geom_segment(aes(xend = proportion, x = proportion_pop, y = level, yend = level),
               color = 'black', 
               alpha = 0.5) + 
  geom_point(size = 1.5, aes(x = proportion, y = level, color = "Sample")) +  # Color inside aes -->
  geom_point(size = 1.5, aes(x = proportion_pop, y = level, color = "Population")) +  # Color inside aes -->
  facet_grid(variable ~ health_zone, scales = 'free_y', space = 'free_y', switch = 'y') + 
  labs(x = "Proportion", y = "") + 
  scale_color_manual(name = "", values = c("Sample" = "black", 'Population' = "red")) + 
  scale_x_continuous(labels = scales::percent_format(), n.breaks = 10) + 
  theme_bw() + 
  theme(legend.position = c(1, 1), # Top-right corner -->
        legend.justification = c("right", "top"), # Anchor the legend at its top-right -->
        legend.direction = 'vertical', 
        legend.margin = margin(t = 0, unit = 'cm'), 
        legend.box.background = element_rect(colour = "black"), 
        legend.title = element_blank())

ggsave(plot = margin_comparison_plot, filename = here("figures", "margin_comparison_plot.png"), width = 10, height = 8)

```
