==> code/11_distn_network_size_fig.Rmd <==
---
title: "Distribution of network size figures"
author: Casey Breen
---

```{r}
## library packages 
library(here)
source(here("code", "helpers.R"))
```



```{r}
## read in survey df
quota_sample_survey <- read_csv(here("data", "data_clean", "quota_sample_survey.csv")) # %>% filter(health_zone != "Nyunzu")

quota_sample_survey <- quota_sample_survey
```


```{r}
quota_sample_survey <-  quota_sample_survey %>% 
  mutate(
    a_total_kin = a_male_child_total + a_female_child_total +
                  a_male_grandchild_total + a_female_grandchild_total +
                  a_sisters_total + a_brothers_total +
                  a_male_cousins_total + a_female_cousins_total +
                  a_parents_total + a_aunts_total +
                  a_uncles_total + a_grandfathers_total + a_grandmothers_total,
                  
    b_total_kin = b_male_child_total + b_female_child_total +
                  b_male_grandchild_total + b_female_grandchild_total +
                  b_sisters_total + b_brothers_total +
                  b_male_cousins_total + b_female_cousins_total +
                  b_parents_total + b_aunts_total +
                  b_uncles_total + b_grandfathers_total + b_grandmothers_total) %>% 
  mutate(a_num_deaths_neighbours = a_num_deaths_hh_neighbours - a_hh_deaths, 
         b_num_deaths_neighbours = b_num_deaths_hh_neighbours - b_hh_deaths)
```


```{r}
quota_sample_survey <- quota_sample_survey %>% 
  mutate(num_total_neighbor = num_total_hh_neighbour - num_total_hh)

neighbor_sample <- quota_sample_survey %>% 
  ggplot(aes(x = num_total_neighbor)) + 
  geom_histogram(binwidth = 2, color = "black", fill = "grey") + 
  theme_cowplot() + 
  xlim(0, 100) + 
  geom_vline(xintercept = mean(quota_sample_survey$num_total_neighbor), color = "blue", linetype = "dashed", linewidth = 1) +
  labs(x = "Neighbor network size", 
       y = "n") + 
  annotate("text", x = 75, y = 200, label = paste0("Mean = ", round(mean(quota_sample_survey$num_total_neighbor), 1)), color = "blue") + 
    ylim(0, 720) 

hh_sample <- quota_sample_survey %>% 
  ggplot(aes(x = num_total_hh)) + 
  geom_histogram(binwidth = 2, color = "black", fill = "grey") +  
  theme_cowplot() + 
  xlim(0, 100) + 
  geom_vline(xintercept = mean(quota_sample_survey$num_total_hh), color = "blue", linetype = "dashed", linewidth = 1) +
    annotate("text", x = 75, y = 500, label = paste0("Mean = ", round(mean(quota_sample_survey$num_total_hh), 1)), color = "blue") +  # Corrected annotation function
  labs(x = "Household size",
       y = "n") + 
    ylim(0, 720) 

kin_sample <- quota_sample_survey %>% 
  ggplot(aes(x = num_total_kin)) + 
  geom_histogram(binwidth = 2, color = "black", fill = "grey") +  
  theme_cowplot() + 
  xlim(0, 100) + 
  geom_vline(xintercept = mean(quota_sample_survey$num_total_kin), color = "blue", linetype = "dashed", linewidth = 1) +
      annotate("text", x = 75, y = 200, label = paste0("Mean = ", round(mean(quota_sample_survey$num_total_kin), 1)), color = "blue") +  # Corrected annotation function
  labs(x = "Kin network size",
       y = "n") + 
    ylim(0, 720) 

plot_network_sizes <- plot_grid(hh_sample, kin_sample, neighbor_sample, ncol = 1)
```



```{r}
# Calculate means and standard errors
plot_data <- quota_sample_survey %>% 
  filter(!is.na(num_deaths_hh)) %>% 
  summarize(
    household_mean = mean(num_deaths_hh, na.rm = TRUE),
    household_se = sd(num_deaths_hh, na.rm = TRUE) / sqrt(n()),
    kin_mean = mean(num_deaths_kin, na.rm = TRUE),
    kin_se = sd(num_deaths_kin, na.rm = TRUE) / sqrt(n()),
    neighbor_mean = mean(num_deaths_neighbours, na.rm = TRUE),
    neighbor_se = sd(num_deaths_neighbours, na.rm = TRUE) / sqrt(n())
  ) %>% 
  pivot_longer(cols = everything(), 
               names_to = c("type", ".value"), 
               names_pattern = "(.*)_(.*)") %>% 
  mutate(type = factor(type, levels = c("neighbor", "kin", "household")))  # Arrange the y-axis order

# Plot with error bars using geom_pointrange
plot_death_reports <- ggplot(plot_data, aes(y = type, x = mean)) + 
  geom_pointrange(aes(xmin = mean - 1.96 * se, xmax = mean + 1.96 * se), size = 0.8, color = "darkred") +  # Add error bars
  xlim(0, 0.15) + 
  labs(x = "Avg. deaths per interview", y = "") + 
  theme_cowplot()
```


```{r}
plot_combined <- plot_grid(plot_network_sizes, plot_death_reports)

ggsave(plot_combined, filename = here("figures", "fig3_network_sizes_plot.png"), width = 7, height = 5)
```
